+++ examples/djangoapp/chat/views.py
@@ -1,10 +1,10 @@
-import json
 import time
 
 from pulsar import is_failure, get_actor
 from pulsar.apps import ws, pubsub
 from pulsar.utils.structures import AttributeDictionary
 from pulsar.utils.log import lazyproperty
+from pulsar.utils.system import json
 
 
 def home(request):
@@ -32,8 +32,10 @@ class Chat(ws.WS):
 
     def pubsub(self, websocket):
         if not self._pubsub:
-            # the ``pulsar.cfg`` is injected by the pulser server into
-            # the wsgi environ
+            # ``pulsar.cfg`` is injected by the pulsar server into
+            # the wsgi environ. Here we pick up the name of the wsgi
+            # application running the server. This is **only** needed by the
+            # test suite which tests several servers/clients at once.
             name = websocket.handshake.environ['pulsar.cfg'].name
             self._pubsub = pubsub.PubSub(name=name)
             self._pubsub.subscribe('webchat')
@@ -42,7 +44,7 @@ class Chat(ws.WS):
     def on_open(self, websocket):
         '''A new websocket connection is established.
 
-        Add connection to the set of clients listening for messages.
+        Add it to the set of clients listening for messages.
         '''
         self.pubsub(websocket).add_client(Client(websocket))
 
@@ -57,7 +59,7 @@ class Chat(ws.WS):
                     lines.append(l)
             msg = ' '.join(lines)
             if msg:
-                user = websocket.handshake.get('django.cache').user
+                user = websocket.handshake.get('django.user')
                 if user.is_authenticated():
                     user = user.username
                 else:
@@ -67,22 +69,22 @@ class Chat(ws.WS):
 
 
 class middleware(object):
-    '''Middleware for serving the Chat websocket.'''
+    '''Django middleware for serving the Chat websocket.'''
     def __init__(self):
         self._web_socket = ws.WebSocket('/message', Chat())
 
     def process_request(self, request):
         from django.http import HttpResponse
-        data = AttributeDictionary(request.__dict__)
-        environ = data.pop('environ')
-        environ['django.cache'] = data
-        response = self._web_socket(environ, None)
+        environ = request.META
+        environ['django.user'] = request.user
+        response = self._web_socket(environ)
         if response is not None:
+            # we have a response, this is the websocket upgrade.
             # Convert to django response
-            if is_failure(response):
-                response.throw()
             resp = HttpResponse(status=response.status_code,
                                 content_type=response.content_type)
             for header, value in response.headers:
                 resp[header] = value
             return resp
+        else:
+            environ.pop('django.user')